back-end/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.9.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
pymysql>=1.0.2
cryptography>=36.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0